# Conflicts:
#	src/main/java/com/Ace/sem/App.java

+++ src/main/java/com/Ace/sem/City.java
@@ -15,9 +15,9 @@ public class City {
     private String Continents;
     private String Region;
 
-    private int totalPopulation;
-    private int totalCitiesPopulation;
-    private int totalNotCitiesPopulation;
+    private long totalPopulation;
+    private long totalCitiesPopulation;
+    private long totalNotCitiesPopulation;
 
     // Getter method for CityName
     public String getCityName() {
@@ -90,31 +90,31 @@ public class City {
     }
 
     // Getter method for Total Population
-    public int getTotalPopulation() {
+    public long getTotalPopulation() {
         return totalPopulation;
     }
 
     // Setter method for Total Population
-    public void setTotalPopulation(int totalPopulation) {
+    public void setTotalPopulation(long totalPopulation) {
         this.totalPopulation = totalPopulation;
     }
 
     // Getter method for Total Population
-    public int getTotalCitiesPopulation() {
+    public long getTotalCitiesPopulation() {
         return totalCitiesPopulation;
     }
 
     // Setter method for Total Population
-    public void setTotalCitiesPopulation(int totalCitiesPopulation) {
+    public void setTotalCitiesPopulation(long totalCitiesPopulation) {
         this.totalCitiesPopulation = totalCitiesPopulation;
     }
     // Getter method for Total Population
-    public int getTotalNotCitiesPopulation() {
+    public long getTotalNotCitiesPopulation() {
         return totalNotCitiesPopulation;
     }
 
     // Setter method for Total Population
-    public void setTotalNotCitiesPopulation(int totalNotCitiesPopulation) {
+    public void setTotalNotCitiesPopulation(long totalNotCitiesPopulation) {
         this.totalNotCitiesPopulation = totalNotCitiesPopulation;
     }
 }
